pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = ["MightContainNuts <mightcontainnuts@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
flask-wtf = "^1.2.2"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
pre-commit = "^4.1.0"
flask-bootstrap = "^3.3.7.1"
flask-migrate = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
